+++ tests/integration/mod.rs
@@ -1,6 +1,7 @@
 mod bash_hook;
 mod build;
 mod code_mappings;
+mod config;
 mod debug_files;
 mod deploys;
 mod events;